refactor: update to check publisher and add throw error at firebase_service.js and update method of  handling error in Participant.js

+++ src/services/firebase_service.js
@@ -54,10 +54,10 @@ class FirebaseService {
             };
             return user;
         }).catch((error) => {
-            // TODO: 여기는 어떻게 처리해야할지 고민...
             const errorCode = error.code;
             const errorMessage = error.message;
             console.log(errorCode, errorMessage);
+            throw new Error("로그인 정보가 잘못됬습니다.");
         });
 
         return user;
@@ -95,7 +95,7 @@ class FirebaseService {
 
             return userInfo.data();
         } else {
-            return null;
+            throw new Error("유저 정보가 존재하지 않습니다.");
         }
     }
 
@@ -163,7 +163,18 @@ class FirebaseService {
      * @param {string} docId 
      */
     async deleteMeeting(docId) {
-        await deleteDoc(doc(this.db, "meetings", docId));
+        const user = await this.getUserInfo();
+        
+        if (sessionStorage.getItem("login flag") === "true") {
+            // 게시자가 유저와 같은지 비교
+            const publisherSnapshot = await getDocs(collection(this.db, 'meetings', docId, 'publisher'));
+            const publisher = publisherSnapshot.docs.map(d => d.data())[0];
+            if (publisher.email === user.email) {
+                await deleteDoc(doc(this.db, "meetings", docId));
+            } else {
+                throw new Error("삭제 권한이 없습니다.");
+            }
+        }
     }
 
     /**
@@ -197,7 +208,6 @@ class FirebaseService {
                 console.log("Transaction failed: ", e);
             }
         } else {
-            // TODO: 로그인 안했을 때 처리 추가
             console.log("로그인 해주세요");
         }
     }
@@ -215,12 +225,12 @@ class FirebaseService {
             
             // email로 중복 참가 방지
             for (let i = 0; i < participants.length; i++) {
-                if (participants[i]['email'] === user['email']) return -1;
+                if (participants[i]['email'] === user['email']) throw new Error("이미 참가한 모임입니다.");
             }
             await addDoc(collection(this.db, `meetings/${docId}/participants`), user);
         } else {
-            // TODO: 로그인 안했을 때 처리 추가
             console.log("로그인 해주세요");
+            throw new Error("로그인 정보가 존재하지 않습니다.");
         }
     }
 
@@ -236,11 +246,15 @@ class FirebaseService {
             const q = query(collection(this.db, `meetings/${docId}/participants`), where("email", "==", user['email']));
             const participantsDoc = await getDocs(q);
             participantsDoc.docs.forEach(async (doc) => {
-                await deleteDoc(doc.ref);
+                if (doc.data().email === user.email)  {
+                    await deleteDoc(doc.ref);
+                } else {
+                    throw new Error("접근 권한이 없습니다.");
+                }
             });
         } else {
-            // TODO: 로그인 안했을 때 처리 추가
             console.log("로그인을 해주세요");
+            throw new Error("로그인 정보가 존재하지 않습니다.");
         }
     }
 }
